# agenthub-cli/docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agenthub",
    "description": "Home Agent Hub. Approved shared context, daily worklogs, and alignment snapshots. Authenticated REST and MCP. Not an LLM.",
    "version": "1.3.0"
  },
  "paths": {
    "/api/v1/context": {
      "get": {
        "summary": "Get Context",
        "operationId": "get_context_api_v1_context_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/grants": {
      "post": {
        "summary": "Put Grant",
        "operationId": "put_grant_api_v1_grants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/capabilities": {
      "get": {
        "summary": "Get Capabilities",
        "operationId": "get_capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/publish-requests/{request_id}/approve": {
      "post": {
        "summary": "Approve Publish",
        "operationId": "approve_publish_api_v1_publish_requests__request_id__approve_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "summary": "Api Me",
        "operationId": "api_me_api_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/uploads": {
      "post": {
        "summary": "Post Upload",
        "operationId": "post_upload_api_v1_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDecl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/uploads/{upload_id}/content": {
      "put": {
        "summary": "Put Upload",
        "operationId": "put_upload_api_v1_uploads__upload_id__content_put",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/publish/plans": {
      "post": {
        "summary": "Post Plan",
        "operationId": "post_plan_api_v1_publish_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/publish/requests": {
      "post": {
        "summary": "Post Plan Request",
        "operationId": "post_plan_request_api_v1_publish_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanReqIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/publish/requests/{request_id}": {
      "get": {
        "summary": "Get Plan Request",
        "operationId": "get_plan_request_api_v1_publish_requests__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要有效 Bearer token 或会话"
          },
          "403": {
            "description": "已认证但无权（写权限不足、仅本人资料或已撤销）"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GrantIn": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": ""
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "project_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Project Ids"
          },
          "scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scopes"
          },
          "share_classes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Classes"
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "agent_id"
        ],
        "title": "GrantIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanReqIn": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "manifest_hash": {
            "type": "string",
            "title": "Manifest Hash"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "manifest_hash"
        ],
        "title": "PlanReqIn"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}